vehicle/viss/0210-subscribe-error.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Signal Server Spec Test: get method tests</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-helper.js"></script>

<h3> Subscribe Test: Receive subscribeErrorResponse if invalid path is specified.</h3>
<h3> Success Condition: SubscribeErrorResponse is returned.</h3>
<h3> Subscribe Signal.Drivetrain.Transmission.InvalidPath</h3>

<h3>
<div id="log"></div>
</h3>

<script>
var t = async_test("Subscribe Test: Receive subscribeErrorResponse");

var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

var NOTIFICATION_COUNT  = 3;

vehicle.onopen = function() {
  // request a signal
  var reqId = "reqid-0210";
  var path = "Signal.Drivetrain.Transmission.InvalidPath";
  var notifCount = 0;
  var phase = PH_START;
  var subId = null;

  // process messages from the server
  vehicle.onmessage = function(event) {
      var msg = JSON.parse(event.data);

      // subscirbe has sent
      if (phase == PH_START) {
        // nothing to do
      } else if (phase == PH_SENT_SUBSCRIBE) {
        t.step_func(function() {
          // wait for subscribe response
          var res = isSubscribeErrorResponse(reqId, msg);
          if (res) {
            assert_true(res, "subscribeErrorResponse received.");
            helper_terminate_success("subscribeErrorResponse received.");
          } else if (isSubscribeSuccessResponse(reqId, msg)) {
            subId = msg.subscriptionId

            // Test failed. Send unsubscribe
            addLogMessage("Send unsubscribe request");
            vehicle.send('{"action":"unsubscribe"'
                          +' ,"subscriptionId":"'+subId+'"'
                          +' ,"requestId":"'+reqId+'"'
                          +'}');
            phase = PH_SENT_UNSUBSCRIBE;
            helper_terminate_failure("subscribeSuccessResponse received.");
          }
        })();
      } else if (phase == PH_SENT_UNSUBSCRIBE) {
        t.step_func(function() {
          // wait for unsubscribe response
          var res = isUnsubscribeSuccessResponse(reqId, subId, msg);

          if (res) {
            // unsubscibe success
            phase = PH_UNSUBSCRIBED;
            helper_terminate_normal("unsubscribeSuccessResponse received");
          } else if (isUnsubscribeErrorResponse(reqId, subId, msg)) {
            // unsubscibe error
            // Test is already finished. Terminate anyway
            helper_terminate_normal("unsubscribeErrorResponse received");
          }
        })();
      } else if (phase == PH_UNSUBSCRIBED) {
        // nothing to do
      }
  }

  // in case server doesn't return response
  t.step_timeout(function() {
    if (phase < PH_SUBSCRIBED) {
      helper_terminate_failure("Finish by timeout");
    }
  }, TIME_OUT_TIME);

  addLogMessage("Send subscribe request: path="+path);
  vehicle.send('{"action":"subscribe","path":"'+ path +'" ,"requestId":"'+reqId+'"}');
  phase = PH_SENT_SUBSCRIBE;
}
</script>
